Move code for getting list of custom bootstrap actions to utils
